chore: move validations model into the validations service
